+++ src/dartfx/rdf/__about__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2024-present kulnor <pascal.heus@gmail.com>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.1.2"
+__version__ = "0.2.0"